blog/oauth2.py
from datetime import datetime, timedelta
from typing import Annotated
from . import JWTtoken
from fastapi import Depends, FastAPI, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from jose import JWTError, jwt
from passlib.context import CryptContext
from pydantic import BaseModel

oauth2_schema = OAuth2PasswordBearer(tokenUrl="Login")#the URL where i chould fetch the Token 

async def get_current_user(token: Annotated[str, Depends(oauth2_schema)]):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    return JWTtoken.verify_Token(token,credentials_exception)